Hypershift: PerformanceProfile adaptation
Performance Profile Controller need to be adapted to properly work on
Hypershift deployments.

When in hypershift deployment Performance Profile controller (PPC) will
not directly reconcile PerformanceProfile objects.
PPC should instead reconcile ConfigMaps in the controller namespace
with a specfic label.
That is because Hypershift Operator (NodePool controller) read
PerformanceProfile objects and embed them in ConfigMaps with that
specific label.

Signed-off-by: Jose Luis Ojosnegros Manchón <jojosneg@redhat.com>